+++ src/info/info.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PairsModule } from 'src/pairs/pairs.module';
 import { PrismaModule } from 'src/prisma/prisma.module';
-import { UtilsModule } from 'src/Utils/utils.module';
+import { UtilsModule } from 'src/utilsModule/utils.module';
 import { InfoController } from './info.controller';
 import { InfoService } from './info.service';
 @Module({
