+++ src/utilsModule/dto/fetchPaths.dto.ts
